move custom checkbox component to inputs
