+++ presenter/src/components/inputs/ContactForm.vue
@@ -113,11 +113,11 @@
 import { useMutation } from '@vue/apollo-composable'
 import { ref } from 'vue'
 
-import DataPrivacyCheckbox from '#components/controls/DataPrivacyCheckbox.vue'
 import AnchorLink from '#components/nav/AnchorLink.vue'
 import { createContactFormMutation } from '#mutations/createContactForm'
 import { messageRules } from '#src/validation/validation'
 
+import DataPrivacyCheckbox from './DataPrivacyCheckbox.vue'
 import EmailInput from './EmailInput.vue'
 import MainButton from './MainButton.vue'
 import NameInput from './NameInput.vue'
