+++ presenter/src/components/inputs/DataPrivacyCheckbox.test.ts
