+++ presenter/src/components/inputs/DataPrivacyCheckbox.vue
