+++ presenter/src/components/inputs/NewsletterForm.vue
@@ -78,10 +78,10 @@
 import { useMutation } from '@vue/apollo-composable'
 import { ref } from 'vue'
 
-import DataPrivacyCheckbox from '#components/controls/DataPrivacyCheckbox.vue'
 import AnchorLink from '#components/nav/AnchorLink.vue'
 import { subscribeToNewsletterMutation } from '#mutations/subscribeToNewsletterMutation'
 
+import DataPrivacyCheckbox from './DataPrivacyCheckbox.vue'
 import EmailInput from './EmailInput.vue'
 import MainButton from './MainButton.vue'
 import NameInput from './NameInput.vue'
